Moved files from middleware to utils.
